signatures: Move Ed25519 types under ed25519 module
This will allow to have a better organization if new algorithms are
added.
