+++ crates/ruma-signatures/src/lib.rs
@@ -62,16 +62,18 @@
 pub use ruma_common::{IdParseError, SigningKeyAlgorithm};
 
 pub use self::{
+    ed25519::{Ed25519KeyPair, Ed25519KeyPairParseError, Ed25519VerificationError},
     error::{JsonError, VerificationError},
     functions::{
         canonical_json, content_hash, hash_and_sign_event, reference_hash, sign_json,
         verify_canonical_json_bytes, verify_event, verify_json,
     },
-    keys::{Ed25519KeyPair, Ed25519KeyPairParseError, KeyPair, PublicKeyMap, PublicKeySet},
+    keys::{KeyPair, PublicKeyMap, PublicKeySet},
     signatures::Signature,
-    verification::{Ed25519VerificationError, Verified},
+    verification::Verified,
 };
 
+mod ed25519;
 mod error;
 mod functions;
 mod keys;
